refactor: unify useChatLogic hook
