+++ src/hooks/useChat.ts
@@ -6,7 +6,7 @@ import { useToast } from "@/hooks/use-toast";
 import type { ChatMessage, Conversation, ChatMessageContentPart } from '@/types';
 import { generateChatTitle } from '@/ai/flows/generate-chat-title';
 import { getPollinationsChatCompletion, type PollinationsChatInput } from '@/ai/flows/pollinations-chat-flow';
-import { DEFAULT_POLLINATIONS_MODEL_ID, DEFAULT_RESPONSE_STYLE_NAME, AVAILABLE_RESPONSE_STYLES, AVAILABLE_POLLINATIONS_MODELS } from '@/config/chat-options';
+import { DEFAULT_POLLINATIONS_MODEL_ID, DEFAULT_RESPONSE_STYLE_NAME, AVAILABLE_RESPONSE_STYLES, AVAILABLE_POLLINATIONS_MODELS, AVAILABLE_TTS_VOICES } from '@/config/chat-options';
 import { textToSpeech } from '@/ai/flows/tts-flow';
 import { speechToText } from '@/ai/flows/stt-flow';
 
@@ -16,26 +16,36 @@ export interface UseChatLogicProps {
   onConversationStarted?: () => void;
 }
 
+// This function contains the entire logic of the original useChat hook
 export function useChatLogic({ userDisplayName, customSystemPrompt, onConversationStarted }: UseChatLogicProps) {
     const [activeConversation, setActiveConversation] = useState<Conversation | null>(null);
     const [allConversations, setAllConversations] = useState<Conversation[]>([]);
     const [currentMessages, setCurrentMessages] = useState<ChatMessage[]>([]);
     const [isAiResponding, setIsAiResponding] = useState(false);
+    
     const [isInitialLoadComplete, setIsInitialLoadComplete] = useState(false);
+  
     const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
     const [chatToDeleteId, setChatToDeleteId] = useState<string | null>(null);
+    
     const [isEditTitleDialogOpen, setIsEditTitleDialogOpen] = useState(false);
     const [chatToEditId, setChatToEditId] = useState<string | null>(null);
     const [editingTitle, setEditingTitle] = useState('');
+    
     const [isImageMode, setIsImageMode] = useState(false);
     const [isHistoryPanelOpen, setIsHistoryPanelOpen] = useState(false);
+  
     const [playingMessageId, setPlayingMessageId] = useState<string | null>(null);
+    const audioRef = useRef<HTMLAudioElement | null>(null);
     const [isRecording, setIsRecording] = useState(false);
     const mediaRecorderRef = useRef<MediaRecorder | null>(null);
     const audioChunksRef = useRef<Blob[]>([]);
     const [chatInputValue, setChatInputValue] = useState('');
-    const { toast } = useToast();
 
+    const [selectedVoice, setSelectedVoice] = useState<string>(AVAILABLE_TTS_VOICES[0].id);
+  
+    const { toast } = useToast();
+  
     const loadConversations = useCallback((loadedConversationsRaw: any) => {
       let loadedConversations: Conversation[] = [];
       if (Array.isArray(loadedConversationsRaw)) {
@@ -45,7 +55,7 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
                   id: conv.id || crypto.randomUUID(),
                   createdAt: new Date(conv.createdAt),
                   messages: (conv.messages || []).map((msg: any) => ({ ...msg, id: msg.id || crypto.randomUUID(), timestamp: new Date(msg.timestamp) })),
-              })).filter((conv: Conversation) => !isNaN(conv.createdAt.getTime()) && conv.toolType && (conv.messages || []).length > 0); // Filter out conversations without a toolType and empty message arrays
+              })).filter((conv: Conversation) => !isNaN(conv.createdAt.getTime()) && conv.toolType === 'long language loops' && conv.messages.some(msg => msg.role === 'user' || msg.role === 'assistant'));
               setAllConversations(loadedConversations.sort((a,b) => b.createdAt.getTime() - a.createdAt.getTime()));
           } catch(e) {
               console.error("Failed to parse conversations", e);
@@ -54,24 +64,64 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
       setIsInitialLoadComplete(true);
       return loadedConversations;
     }, []);
-
+  
     useEffect(() => {
       if (!isInitialLoadComplete) {
         return;
       }
-      const conversationsToStore = allConversations
-          .filter(conv => conv.toolType && (conv.messages || []).some(msg => msg.role === 'user' || msg.role === 'assistant')) // Ensure toolType exists and filter by messages
+      try {
+        const conversationsToStore = allConversations
+          .filter(conv => conv.toolType === 'long language loops' && conv.messages.some(msg => msg.role === 'user' || msg.role === 'assistant'))
           .map(conv => {
-              const { uploadedFile, ...storableConv } = conv;
-              return storableConv;
+            // Create a deep copy to avoid mutating the active state
+            const storableConv = JSON.parse(JSON.stringify(conv));
+            
+            // Remove non-serializable File object
+            delete storableConv.uploadedFile;
+    
+            // Process messages to handle large data URIs
+            storableConv.messages = storableConv.messages.map((msg: ChatMessage) => {
+              if (Array.isArray(msg.content)) {
+                msg.content = msg.content.map(part => {
+                  if (part.type === 'image_url' && part.image_url.url.startsWith('data:image')) {
+                    // Replace large data URI with a placeholder to prevent storage overflow
+                    return {
+                      ...part,
+                      image_url: {
+                        ...part.image_url,
+                        url: 'https://placehold.co/512x512.png?text=Image+History+Disabled',
+                        altText: 'Image removed from history to save space.',
+                        isGenerated: false,
+                        isUploaded: false,
+                      }
+                    };
+                  }
+                  return part;
+                });
+              }
+              return msg;
+            });
+    
+            // Also handle the top-level uploadedFilePreview
+            if (storableConv.uploadedFilePreview && storableConv.uploadedFilePreview.startsWith('data:image')) {
+              storableConv.uploadedFilePreview = null;
+            }
+    
+            return storableConv;
           });
-      if (conversationsToStore.length > 0) {
-          localStorage.setItem('chatConversations', JSON.stringify(conversationsToStore));
-      } else {
-          localStorage.removeItem('chatConversations');
+          
+        if (conversationsToStore.length > 0) {
+            localStorage.setItem('chatConversations', JSON.stringify(conversationsToStore));
+        } else {
+            localStorage.removeItem('chatConversations');
+        }
+      } catch (error) {
+          console.error("Error saving conversations to localStorage (might be full):", error);
+          // Don't toast here as it can be annoying, but the error is logged.
+          // The user already sees the crash boundary if this fails during a write operation.
       }
     }, [allConversations, isInitialLoadComplete]);
-
+  
     const updateActiveConversationState = useCallback((updates: Partial<Conversation>) => {
       setActiveConversation(prevActive => {
         if (!prevActive) return null;
@@ -79,12 +129,12 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
         setAllConversations(prevAllConvs => prevAllConvs.map(c => (c.id === prevActive.id ? updatedConv : c)));
         return updatedConv;
       });
-
-      if (updates.hasOwnProperty('isImageMode')) {
+  
+      if (updates.hasOwnProperty('isImageMode')) { 
           setIsImageMode(updates.isImageMode || false);
       }
     }, []);
-
+  
     useEffect(() => {
       if (activeConversation) {
         setCurrentMessages(activeConversation.messages);
@@ -94,13 +144,13 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
         setIsImageMode(false);
       }
     }, [activeConversation]);
-
+  
     const updateConversationTitle = useCallback(async (conversationId: string, messagesForTitleGen: ChatMessage[]) => {
       const convToUpdate = allConversations.find(c => c.id === conversationId);
-      if (!convToUpdate || !convToUpdate.toolType) return; // Ensure toolType exists
-
+      if (!convToUpdate || convToUpdate.toolType !== 'long language loops') return;
+  
       const isDefaultTitle = convToUpdate.title === "default.long.language.loop" || convToUpdate.title.toLowerCase().startsWith("new ") || convToUpdate.title === "Chat";
-
+  
       if (messagesForTitleGen.length >= 1 && messagesForTitleGen.length < 5 && isDefaultTitle) {
         const relevantText = messagesForTitleGen.map(msg => typeof msg.content === 'string' ? `${msg.role}: ${msg.content}` : `${msg.role}: ${msg.content.find(p => p.type === 'text')?.text || ''}`).filter(Boolean).slice(0, 3).join('\n\n');
         if (relevantText) {
@@ -114,20 +164,20 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
         }
       }
     }, [allConversations, activeConversation?.id]);
-
+    
     const sendMessage = useCallback(async (
       _messageText: string,
       options: { isImageModeIntent?: boolean; } = {}
     ) => {
       const messageText = chatInputValue.trim();
-      if (!activeConversation || !activeConversation.toolType || (!messageText && !activeConversation.uploadedFile)) return; // Ensure toolType exists
-
+      if (!activeConversation || activeConversation.toolType !== 'long language loops' || (!messageText && !activeConversation.uploadedFile)) return;
+  
       const { selectedModelId, selectedResponseStyleName, messages, uploadedFile, uploadedFilePreview } = activeConversation;
       const currentModel = AVAILABLE_POLLINATIONS_MODELS.find(m => m.id === selectedModelId) || AVAILABLE_POLLINATIONS_MODELS[0];
-
+      
       let effectiveSystemPrompt = '';
       const basicStylePrompt = (AVAILABLE_RESPONSE_STYLES.find(s => s.name === 'Basic') || AVAILABLE_RESPONSE_STYLES[0]).systemPrompt;
-
+  
       if (selectedResponseStyleName === "User's Default") {
           if (customSystemPrompt && customSystemPrompt.trim()) {
               effectiveSystemPrompt = customSystemPrompt.replace(/{userDisplayName}/gi, userDisplayName || "User");
@@ -138,19 +188,19 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
           const selectedStyle = AVAILABLE_RESPONSE_STYLES.find(s => s.name === selectedResponseStyleName);
           effectiveSystemPrompt = selectedStyle ? selectedStyle.systemPrompt : basicStylePrompt;
       }
-
+  
       setIsAiResponding(true);
       setChatInputValue('');
       const convId = activeConversation.id;
       const isImagePrompt = options.isImageModeIntent || false;
       const isFileUpload = !!uploadedFile && !isImagePrompt;
-
+  
       if (isFileUpload && !currentModel.vision) {
         toast({ title: "Model Incompatibility", description: `Model '${currentModel.name}' doesn't support images.`, variant: "destructive" });
         setIsAiResponding(false);
         return;
       }
-
+  
       let userMessageContent: string | ChatMessageContentPart[] = messageText;
       if (isFileUpload && uploadedFilePreview) {
         userMessageContent = [
@@ -158,13 +208,13 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
           { type: 'image_url', image_url: { url: uploadedFilePreview, altText: uploadedFile.name, isUploaded: true } }
         ];
       }
-
-      const userMessage: ChatMessage = { id: crypto.randomUUID(), role: 'user', content: userMessageContent, timestamp: new Date(), toolType: activeConversation.toolType }; // Use activeConversation.toolType
+  
+      const userMessage: ChatMessage = { id: crypto.randomUUID(), role: 'user', content: userMessageContent, timestamp: new Date(), toolType: 'long language loops' };
       const messagesForApi = [...messages, userMessage];
       const updatedMessagesForState = isImagePrompt ? messages : [...messages, userMessage];
-
+      
       updateActiveConversationState({ messages: updatedMessagesForState });
-
+  
       let aiResponseContent: string | ChatMessageContentPart[] | null = null;
       try {
           if (isImagePrompt && messageText) {
@@ -189,20 +239,20 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
           toast({ title: "AI Error", description: errorMessage, variant: "destructive" });
           aiResponseContent = `Sorry, an error occurred: ${errorMessage}`;
       }
-
+  
       if (aiResponseContent !== null) {
-        const aiMessage: ChatMessage = { id: crypto.randomUUID(), role: 'assistant', content: aiResponseContent, timestamp: new Date(), toolType: activeConversation.toolType }; // Use activeConversation.toolType
+        const aiMessage: ChatMessage = { id: crypto.randomUUID(), role: 'assistant', content: aiResponseContent, timestamp: new Date(), toolType: 'long language loops' };
         const finalMessages = [...updatedMessagesForState, aiMessage];
         updateActiveConversationState({ messages: finalMessages });
         updateConversationTitle(convId, finalMessages);
       }
-
+      
       if (isImagePrompt || isFileUpload) {
           updateActiveConversationState({ isImageMode: false, uploadedFile: null, uploadedFilePreview: null });
       }
       setIsAiResponding(false);
     }, [activeConversation, customSystemPrompt, userDisplayName, toast, updateActiveConversationState, updateConversationTitle, chatInputValue]);
-
+  
     const selectChat = useCallback((conversationId: string | null) => {
       if (conversationId === null) {
           setActiveConversation(null);
@@ -210,27 +260,27 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
       }
       const conversationToSelect = allConversations.find(c => c.id === conversationId);
       if (!conversationToSelect) return;
-
+  
       setActiveConversation({ ...conversationToSelect, uploadedFile: null });
       onConversationStarted?.();
     }, [allConversations, onConversationStarted]);
-
+    
     const startNewChat = useCallback(() => {
       const newConversation: Conversation = {
         id: crypto.randomUUID(),
         title: "default.long.language.loop",
         messages: [],
         createdAt: new Date(),
-        toolType: 'long language loops', // Keep this for now, but consider making it dynamic
-        isImageMode: false,
-        selectedModelId: DEFAULT_POLLINATIONS_MODEL_ID,
+        toolType: 'long language loops',
+        isImageMode: false, 
+        selectedModelId: DEFAULT_POLLINATIONS_MODEL_ID, 
         selectedResponseStyleName: DEFAULT_RESPONSE_STYLE_NAME,
       };
       setAllConversations(prev => [newConversation, ...prev].sort((a, b) => b.createdAt.getTime() - a.createdAt.getTime()));
       setActiveConversation(newConversation);
       onConversationStarted?.();
     }, [onConversationStarted]);
-
+    
     const requestEditTitle = (conversationId: string) => {
       const convToEdit = allConversations.find(c => c.id === conversationId);
       if (!convToEdit) return;
@@ -238,7 +288,7 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
       setEditingTitle(convToEdit.title);
       setIsEditTitleDialogOpen(true);
     };
-
+    
     const confirmEditTitle = () => {
       if (!chatToEditId || !editingTitle.trim()) {
         toast({ title: "Invalid Title", description: "Title cannot be empty.", variant: "destructive" });
@@ -252,21 +302,21 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
       toast({ title: "Title Updated" });
       setIsEditTitleDialogOpen(false);
     };
-
+  
     const cancelEditTitle = () => setIsEditTitleDialogOpen(false);
-
+  
     const requestDeleteChat = (conversationId: string) => {
       setChatToDeleteId(conversationId);
       setIsDeleteDialogOpen(true);
     };
-
+    
     const deleteChat = (conversationId: string, silent = false) => {
         const wasActive = activeConversation?.id === conversationId;
         const updatedConversations = allConversations.filter(c => c.id !== conversationId);
         setAllConversations(updatedConversations);
-
+  
         if (wasActive) {
-          const nextChat = updatedConversations.find(c => c.toolType); // Find the next chat with any toolType
+          const nextChat = updatedConversations.find(c => c.toolType === 'long language loops');
           if (nextChat) selectChat(nextChat.id);
           else startNewChat();
         }
@@ -274,21 +324,21 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
           toast({ title: "Chat Deleted" });
         }
     }
-
+  
     const confirmDeleteChat = () => {
       if (!chatToDeleteId) return;
       deleteChat(chatToDeleteId);
       setIsDeleteDialogOpen(false);
     };
-
+  
     const cancelDeleteChat = () => setIsDeleteDialogOpen(false);
-
+  
     const toggleImageMode = () => {
       if (!activeConversation) return;
-      const newImageModeState = !isImageMode;
+      const newImageModeState = !isImageMode; 
       updateActiveConversationState({ isImageMode: newImageModeState, uploadedFile: null, uploadedFilePreview: null });
     };
-
+  
     const handleFileSelect = (file: File | null) => {
       if (!activeConversation) return;
       if (file) {
@@ -301,55 +351,87 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
         updateActiveConversationState({ uploadedFile: null, uploadedFilePreview: null });
       }
     };
-
+  
     const clearUploadedImage = () => {
-      if (activeConversation) handleFileSelect(null);
+      if (activeConversation) handleFileSelect(null); 
     }
-
+  
     const handleModelChange = useCallback((modelId: string) => {
       if (activeConversation) updateActiveConversationState({ selectedModelId: modelId });
     }, [activeConversation, updateActiveConversationState]);
-
+  
     const handleStyleChange = useCallback((styleName: string) => {
        if (activeConversation) updateActiveConversationState({ selectedResponseStyleName: styleName });
     }, [activeConversation, updateActiveConversationState]);
 
+    const handleVoiceChange = (voiceId: string) => {
+      setSelectedVoice(voiceId);
+    };
+  
     const toggleHistoryPanel = () => setIsHistoryPanelOpen(prev => !prev);
     const closeHistoryPanel = useCallback(() => setIsHistoryPanelOpen(false), []);
-
-    // ---- HIER IST DER EINZIGE GEÄNDERTE TEIL: ----
+  
     const handlePlayAudio = useCallback(async (text: string, messageId: string) => {
-      if (playingMessageId) return;
-      setPlayingMessageId(messageId);
+      // If the user clicks the button of the currently playing audio, stop it.
+      if (audioRef.current && playingMessageId === messageId) {
+        audioRef.current.pause();
+        audioRef.current = null;
+        setPlayingMessageId(null);
+        return;
+      }
+  
+      // If another audio is playing, stop it before starting the new one.
+      if (audioRef.current) {
+        audioRef.current.pause();
+        audioRef.current = null;
+      }
+  
+      if (!text || !text.trim()) {
+        setPlayingMessageId(null); // Ensure state is reset if there's nothing to play
+        return;
+      }
+      
+      setPlayingMessageId(messageId); // Indicate that we are loading/playing this message
+      
       try {
-        // Zweites Argument z.B. "alloy", "echo", "nova" etc.
-        const { audioDataUri } = await textToSpeech(text, "alloy");
+        const { audioDataUri } = await textToSpeech(text, selectedVoice);
         const audio = new Audio(audioDataUri);
+        audioRef.current = audio;
         audio.play();
+  
         audio.onended = () => {
-          setPlayingMessageId(null);
+          // Only clear state if this is the audio that just finished
+          if (audioRef.current === audio) {
+            audioRef.current = null;
+            setPlayingMessageId(null);
+          }
         };
         audio.onerror = () => {
           toast({ title: "Audio Playback Error", description: "Could not play the generated audio.", variant: "destructive" });
-          setPlayingMessageId(null);
-        };
+          if (audioRef.current === audio) {
+            audioRef.current = null;
+            setPlayingMessageId(null);
+          }
+        }
       } catch (error) {
         console.error("TTS Error:", error);
         toast({ title: "Text-to-Speech Error", description: "Could not generate audio.", variant: "destructive" });
+        audioRef.current = null; // Clean up on error
         setPlayingMessageId(null);
       }
-    }, [playingMessageId, toast]);
-    // -----------------------------------------------
-
+    }, [playingMessageId, toast, selectedVoice]);
+  
     const handleStartRecording = useCallback(async () => {
       if (isRecording) return;
       try {
         const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
         mediaRecorderRef.current = new MediaRecorder(stream);
         audioChunksRef.current = [];
+  
         mediaRecorderRef.current.ondataavailable = (event) => {
           audioChunksRef.current.push(event.data);
         };
+  
         mediaRecorderRef.current.onstop = async () => {
           const audioBlob = new Blob(audioChunksRef.current, { type: 'audio/webm' });
           const reader = new FileReader();
@@ -369,25 +451,23 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
               }
           };
         };
+  
         mediaRecorderRef.current.start();
         setIsRecording(true);
       } catch (error) {
         console.error("Microphone access error:", error);
         toast({ title: "Microphone Access Denied", description: "Please enable microphone permissions in your browser.", variant: "destructive" });
       }
-    }, [isRecording, toast, setChatInputValue, setIsAiResponding]);
-
+    }, [isRecording, toast]);
+  
     const handleStopRecording = useCallback(() => {
       if (mediaRecorderRef.current && isRecording) {
         mediaRecorderRef.current.stop();
-        // Stop all tracks in the stream to release the microphone
-        if (mediaRecorderRef.current.stream) {
-            mediaRecorderRef.current.stream.getTracks().forEach(track => track.stop());
-        }
+        mediaRecorderRef.current.stream.getTracks().forEach(track => track.stop());
         setIsRecording(false);
       }
     }, [isRecording]);
-
+  
     const handleToggleRecording = useCallback(() => {
       if (isRecording) {
         handleStopRecording();
@@ -395,16 +475,80 @@ export function useChatLogic({ userDisplayName, customSystemPrompt, onConversati
         handleStartRecording();
       }
     }, [isRecording, handleStartRecording, handleStopRecording]);
-
+  
+    const handleCopyToClipboard = useCallback((text: string) => {
+      if (!text) return;
+      navigator.clipboard.writeText(text).then(() => {
+        toast({ title: "Copied to Clipboard" });
+      }).catch(err => {
+        console.error("Failed to copy text: ", err);
+        toast({ title: "Copy Failed", variant: "destructive" });
+      });
+    }, [toast]);
+  
+    const regenerateLastResponse = useCallback(async () => {
+      if (isAiResponding || !activeConversation) return;
+  
+      const lastMessageIndex = activeConversation.messages.length - 1;
+      const lastMessage = activeConversation.messages[lastMessageIndex];
+  
+      if (!lastMessage || lastMessage.role !== 'assistant') {
+        toast({ title: "Action Not Available", description: "You can only regenerate the AI's most recent response.", variant: "destructive" });
+        return;
+      }
+  
+      setIsAiResponding(true);
+      const historyForApi = activeConversation.messages.slice(0, lastMessageIndex);
+      updateActiveConversationState({ messages: historyForApi });
+  
+      let aiResponseContent: string | ChatMessageContentPart[] | null = null;
+      try {
+        const { selectedModelId, selectedResponseStyleName } = activeConversation;
+        const currentModel = AVAILABLE_POLLINATIONS_MODELS.find(m => m.id === selectedModelId) || AVAILABLE_POLLINATIONS_MODELS[0];
+        let effectiveSystemPrompt = '';
+        const basicStylePrompt = (AVAILABLE_RESPONSE_STYLES.find(s => s.name === 'Basic') || AVAILABLE_RESPONSE_STYLES[0]).systemPrompt;
+        if (selectedResponseStyleName === "User's Default") {
+          effectiveSystemPrompt = (customSystemPrompt && customSystemPrompt.trim()) ? customSystemPrompt.replace(/{userDisplayName}/gi, userDisplayName || "User") : basicStylePrompt;
+        } else {
+          const selectedStyle = AVAILABLE_RESPONSE_STYLES.find(s => s.name === selectedResponseStyleName);
+          effectiveSystemPrompt = selectedStyle ? selectedStyle.systemPrompt : basicStylePrompt;
+        }
+  
+        const apiInput: PollinationsChatInput = { messages: historyForApi, modelId: currentModel.id, systemPrompt: effectiveSystemPrompt };
+        const result = await getPollinationsChatCompletion(apiInput);
+        aiResponseContent = result.responseText;
+  
+      } catch (error) {
+        const errorMessage = error instanceof Error ? error.message : "An unknown error occurred.";
+        toast({ title: "AI Error", description: errorMessage, variant: "destructive" });
+        aiResponseContent = `Sorry, an error occurred: ${errorMessage}`;
+        updateActiveConversationState({ messages: [...historyForApi, lastMessage] }); // Restore original message on error
+      }
+  
+      if (aiResponseContent !== null) {
+        const aiMessage: ChatMessage = { id: crypto.randomUUID(), role: 'assistant', content: aiResponseContent, timestamp: new Date(), toolType: 'long language loops' };
+        const finalMessages = [...historyForApi, aiMessage];
+        updateActiveConversationState({ messages: finalMessages });
+      }
+      
+      setIsAiResponding(false);
+    }, [isAiResponding, activeConversation, updateActiveConversationState, customSystemPrompt, userDisplayName, toast]);
+  
+  
     return {
       activeConversation, allConversations, currentMessages, isAiResponding, isImageMode,
       isHistoryPanelOpen, isDeleteDialogOpen, isEditTitleDialogOpen, editingTitle,
       isRecording, playingMessageId, chatInputValue,
+      selectedVoice,
       loadConversations, selectChat, startNewChat, deleteChat, sendMessage,
       requestEditTitle, confirmEditTitle, cancelEditTitle, setEditingTitle,
       requestDeleteChat, confirmDeleteChat, cancelDeleteChat, toggleImageMode,
       handleFileSelect, clearUploadedImage, handleModelChange, handleStyleChange,
+      handleVoiceChange,
       toggleHistoryPanel, closeHistoryPanel, handlePlayAudio, handleToggleRecording,
       setChatInputValue,
+      handleCopyToClipboard,
+      regenerateLastResponse,
     };
 }
+
